// src/Features/pastQuestion/enums/pastQuestion.enum.ts
export enum PastQuestionStatus {
  ACTIVE = "ACTIVE",
  DEACTIVE = "DEACTIVE",
}

export enum PastQuestionProcessStatus {
  PENDING = "PENDING",
  PROCESSING = "PROCESSING",
  SUCCESS = "SUCCESS",
  FAILED = "FAILED",
}

export enum PastQuestionInsightScope {
  PAPER = "paper",
  COURSE = "course",
}

export enum PastQuestionChatScope {
  COURSE = "course",
  PAPER = "paper",
  QUESTION = "question",
}